automatically consider global step on training feed dict in TRHG

+++ far_ho/hyper_gradients.py
@@ -302,7 +302,7 @@ class ReverseHG(HyperGradient):
                     _T = 0
                     # print('WARNING, USING TRUNCATED REVERSE without setting a global step')
                 tot_t = utils.solve_int_or_generator(T_or_generator[0])
-                return _t + tot_t*_T
+                return int(_t + tot_t*_T)
             else: return _t
 
         if not online:
